feat(lap-detector-v2): wire quality check and sector computation
Extract computeLapSectors from v1 into server/compute-lap-sectors.ts as a
pure function. v1 now delegates to the extracted function. v2 imports it
and calls assessLapRecording for validity.
