+++ server/lap-detector.ts
@@ -12,9 +12,9 @@
  */
 import type { TelemetryPacket, GameId } from "../shared/types";
 import type { DbAdapter } from "./pipeline-adapters";
-import { extractCurbSegments, recordCurbData, getTrackSectorsByOrdinal, loadSharedTrackMeta } from "../shared/track-data";
+import { extractCurbSegments, recordCurbData } from "../shared/track-data";
 import { assessLapRecording } from "./lap-quality";
-import { tryGetGame } from "../shared/games/registry";
+import { computeLapSectors as computeLapSectorsHelper } from "./compute-lap-sectors";
 import { detectSessionBoundary, detectLapBoundary, detectLapReset } from "./lap-detection";
 
 
@@ -540,99 +540,12 @@ export class LapDetector {
     packets: TelemetryPacket[],
     lapTime: number
   ): Promise<{ s1: number; s2: number; s3: number } | null> {
-    if (!this.currentSession || packets.length < 50) return null;
+    if (!this.currentSession) return null;
     const { trackOrdinal, gameId } = this.currentSession;
-
-    // Resolve sector boundaries
-    const adapter = tryGetGame(gameId);
-    const sharedName = adapter?.getSharedTrackName?.(trackOrdinal);
-    const dbSectors = await this.db.getTrackOutlineSectors(trackOrdinal, gameId);
-    const sharedMeta = sharedName ? loadSharedTrackMeta(sharedName) : null;
-    const gameSectors = gameId ? (sharedMeta as any)?.games?.[gameId]?.sectors : null;
-    const raw = dbSectors ?? gameSectors ?? sharedMeta?.sectors ?? getTrackSectorsByOrdinal(trackOrdinal);
-    const s1End = raw?.s1End ?? 1 / 3;
-    const s2End = raw?.s2End ?? 2 / 3;
-
-    // F1: prefer game-broadcast sector times
-    let s1 = 0, s2 = 0;
-    for (const p of packets) {
-      if ((p.f1?.sector1Time ?? 0) > 0) s1 = p.f1!.sector1Time;
-      if ((p.f1?.sector2Time ?? 0) > 0) s2 = p.f1!.sector2Time;
-    }
-
-    // ACC: use native sector times tracked live during the lap
-    if (s1 === 0 && s2 === 0 && gameId === "acc" && this.accS1 > 0 && this.accS2 > 0) {
-      s1 = this.accS1;
-      s2 = this.accS2;
-    }
-
-    // Fall back to distance-fraction computation
-    if (s1 === 0 || s2 === 0) {
-      const startDist = packets[0].DistanceTraveled;
-      const lapDist = packets[packets.length - 1].DistanceTraveled - startDist;
-      if (lapDist < 100) return null;
-
-      let sector = 0;
-      let sectorStart = packets[0].CurrentLap;
-      s1 = 0;
-      s2 = 0;
-      for (const p of packets) {
-        const frac = (p.DistanceTraveled - startDist) / lapDist;
-        const expected = frac < s1End ? 0 : frac < s2End ? 1 : 2;
-        if (expected > sector) {
-          const t = p.CurrentLap - sectorStart;
-          if (sector === 0) s1 = t;
-          else if (sector === 1) s2 = t;
-          sectorStart = p.CurrentLap;
-          sector = expected;
-        }
-      }
-    }
-
-    if (s1 > 0 && s2 > 0) {
-      const s3 = lapTime - s1 - s2;
-      if (s3 <= 0) {
-        // Native sectors invalid — fall through to distance-fraction fallback
-        s1 = 0;
-        s2 = 0;
-      } else {
-        return { s1, s2, s3 };
-      }
-    }
-
-    // If we get here, native sectors didn't work, try distance-fraction one more time
-    if (s1 === 0 || s2 === 0) {
-      const startDist = packets[0].DistanceTraveled;
-      const lapDist = packets[packets.length - 1].DistanceTraveled - startDist;
-      if (lapDist >= 100) {
-        const s1End = raw?.s1End ?? 1 / 3;
-        const s2End = raw?.s2End ?? 2 / 3;
-
-        let sector = 0;
-        let sectorStart = packets[0].CurrentLap;
-        let s1Retry = 0, s2Retry = 0;
-        for (const p of packets) {
-          const frac = (p.DistanceTraveled - startDist) / lapDist;
-          const expected = frac < s1End ? 0 : frac < s2End ? 1 : 2;
-          if (expected > sector) {
-            const t = p.CurrentLap - sectorStart;
-            if (sector === 0) s1Retry = t;
-            else if (sector === 1) s2Retry = t;
-            sectorStart = p.CurrentLap;
-            sector = expected;
-          }
-        }
-
-        if (s1Retry > 0 && s2Retry > 0) {
-          const s3 = lapTime - s1Retry - s2Retry;
-          if (s3 > 0) {
-            return { s1: s1Retry, s2: s2Retry, s3 };
-          }
-        }
-      }
-    }
-
-    return null;
+    const accLiveSectors = this.accS1 > 0 && this.accS2 > 0
+      ? { s1: this.accS1, s2: this.accS2 }
+      : undefined;
+    return computeLapSectorsHelper(this.db, trackOrdinal, gameId, packets, lapTime, accLiveSectors);
   }
 
   private resetLapState(newLapFirstPacket: TelemetryPacket, seedPackets: TelemetryPacket[] = []): void {
